[#47] refactor: 게시글 상세 페이지 별도 컴포넌트로 분리하여 리팩토링
